improve

+++ r/eae31718-c486-4e40-ad71-c2416c2efd05/libs/Editor_files/jhashtable.js
@@ -34,263 +34,175 @@
  * </code>
  */
 
-gamecore.Hashtable = (function ()
-{
-    var FUNCTION = "function";
+(function () {
+    'use strict';
 
-    var arrayRemoveAt = (typeof Array.prototype.splice == FUNCTION) ?
-        function (arr, idx)
-        {
-            arr.splice(idx, 1);
-        } :
-
-        function (arr, idx)
-        {
-            var itemsAfterDeleted, i, len;
-            if (idx === arr.length - 1)
-            {
-                arr.length = idx;
-            } else
-            {
-                itemsAfterDeleted = arr.slice(idx + 1);
-                arr.length = idx;
-                for (i = 0, len = itemsAfterDeleted.length; i < len; ++i)
-                {
-                    arr[idx + i] = itemsAfterDeleted[i];
-                }
-            }
-        };
-
-    function hashObject(obj)
-    {
-        var hashCode;
-        if (typeof obj == "string")
-        {
+    function hashObject(obj) {
+        if (typeof obj === 'string') {
             return obj;
-        } else if (typeof obj.hashCode == FUNCTION)
-        {
-            // Check the hashCode method really has returned a string
-            hashCode = obj.hashCode();
-            return (typeof hashCode == "string") ? hashCode : hashObject(hashCode);
-        } else if (typeof obj.toString == FUNCTION)
-        {
+        } else if (typeof obj.hashCode === 'function') {
+            return obj.hashCode();
+        } else if (typeof obj.toString === 'function') {
             return obj.toString();
-        } else
-        {
-            try
-            {
+        } else {
+            try {
                 return String(obj);
-            }
-            catch (ex)
-            {
-                // For host objects (such as ActiveObjects in IE) that have no toString() method and throw an error when
-                // passed to String()
+            } catch (ex) {
                 return Object.prototype.toString.call(obj);
             }
         }
     }
 
-    function equals_fixedValueHasEquals(fixedValue, variableValue)
-    {
+    function equals_fixedValueHasEquals(fixedValue, variableValue) {
         return fixedValue.equals(variableValue);
     }
 
-    function equals_fixedValueNoEquals(fixedValue, variableValue)
-    {
-        return (typeof variableValue.equals == FUNCTION) ?
+    function equals_fixedValueNoEquals(fixedValue, variableValue) {
+        return typeof variableValue.equals === 'function' ?
             variableValue.equals(fixedValue) : (fixedValue === variableValue);
     }
 
-    function createKeyValCheck(kvStr)
-    {
-        return function (kv)
-        {
-            if (kv === null)
-            {
-                throw new Error("null is not a valid " + kvStr);
-            } else if (typeof kv == "undefined")
-            {
-                throw new Error(kvStr + " must not be undefined");
-            }
-        };
-    }
-
-    var checkKey = createKeyValCheck("key"), checkValue = createKeyValCheck("value");
-
-    /*----------------------------------------------------------------------------------------------------------------*/
-
-    function Bucket(hash, firstKey, firstValue, equalityFunction)
-    {
+    function Bucket(hash, firstKey, firstValue, equalityFunction) {
         this[0] = hash;
         this.entries = [];
         this.addEntry(firstKey, firstValue);
 
-        if (equalityFunction !== null)
-        {
-            this.getEqualityFunction = function ()
-            {
+        if (equalityFunction !== null) {
+            this.getEqualityFunction = function () {
                 return equalityFunction;
             };
         }
     }
 
-    var EXISTENCE = 0, ENTRY = 1, ENTRY_INDEX_AND_VALUE = 2;
-
-    function createBucketSearcher(mode)
-    {
-        return function (key)
-        {
-            var i = this.entries.length, entry, equals = this.getEqualityFunction(key);
-            while (i--)
-            {
-                entry = this.entries[i];
-                if (equals(key, entry[0]))
-                {
-                    switch (mode)
-                    {
-                        case EXISTENCE:
-                            return true;
-                        case ENTRY:
-                            return entry;
-                        case ENTRY_INDEX_AND_VALUE:
-                            return [ i, entry[1] ];
-                    }
-                }
-            }
-            return false;
-        };
-    }
-
-    function createBucketLister(entryProperty)
-    {
-        return function (aggregatedArr)
-        {
-            var startIndex = aggregatedArr.length;
-            for (var i = 0, len = this.entries.length; i < len; ++i)
-            {
-                aggregatedArr[startIndex + i] = this.entries[i][entryProperty];
-            }
-        };
-    }
-
     Bucket.prototype = {
-        getEqualityFunction:function (searchValue)
-        {
-            return (typeof searchValue.equals == FUNCTION) ? equals_fixedValueHasEquals : equals_fixedValueNoEquals;
+        getEqualityFunction: function (searchValue) {
+            return typeof searchValue.equals === 'function' ?
+                equals_fixedValueHasEquals : equals_fixedValueNoEquals;
         },
 
-        getEntryForKey:createBucketSearcher(ENTRY),
+        getEntryForKey: function (key) {
+            for (var i = 0, len = this.entries.length; i < len; ++i) {
+                if (this.getEqualityFunction(key)(key, this.entries[i][0])) {
+                    return this.entries[i];
+                }
+            }
+            return null;
+        },
 
-        getEntryAndIndexForKey:createBucketSearcher(ENTRY_INDEX_AND_VALUE),
+        getEntryAndIndexForKey: function (key) {
+            for (var i = 0, len = this.entries.length; i < len; ++i) {
+                if (this.getEqualityFunction(key)(key, this.entries[i][0])) {
+                    return [i, this.entries[i][1]];
+                }
+            }
+            return null;
+        },
 
-        removeEntryForKey:function (key)
-        {
+        removeEntryForKey: function (key) {
             var result = this.getEntryAndIndexForKey(key);
-            if (result)
-            {
-                arrayRemoveAt(this.entries, result[0]);
+            if (result) {
+                this.entries.splice(result[0], 1);
                 return result[1];
             }
             return null;
         },
 
-        addEntry:function (key, value)
-        {
-            this.entries[this.entries.length] = [key, value];
+        addEntry: function (key, value) {
+            this.entries.push([key, value]);
         },
 
-        keys:createBucketLister(0),
-
-        values:createBucketLister(1),
-
-        getEntries:function (entries)
-        {
-            var startIndex = entries.length;
-            for (var i = 0, len = this.entries.length; i < len; ++i)
-            {
-                // Clone the entry stored in the bucket before adding to array
-                entries[startIndex + i] = this.entries[i].slice(0);
+        keys: function (aggregatedArr) {
+            if (!aggregatedArr) {
+                aggregatedArr = [];
             }
+            for (var i = 0, len = this.entries.length; i < len; ++i) {
+                aggregatedArr.push(this.entries[i][0]);
+            }
+            return aggregatedArr;
         },
 
-        containsKey:createBucketSearcher(EXISTENCE),
+        values: function (aggregatedArr) {
+            if (!aggregatedArr) {
+                aggregatedArr = [];
+            }
+            for (var i = 0, len = this.entries.length; i < len; ++i) {
+                aggregatedArr.push(this.entries[i][1]);
+            }
+            return aggregatedArr;
+        },
 
-        containsValue:function (value)
-        {
-            var i = this.entries.length;
-            while (i--)
-            {
-                if (value === this.entries[i][1])
-                {
+        getEntries: function (entries) {
+            if (!entries) {
+                entries = [];
+            }
+            for (var i = 0, len = this.entries.length; i < len; ++i) {
+                entries.push(this.entries[i].slice(0));
+            }
+            return entries;
+        },
+
+        containsKey: function (key) {
+            return !!this.getEntryForKey(key);
+        },
+
+        containsValue: function (value) {
+            for (var i = 0, len = this.entries.length; i < len; ++i) {
+                if (value === this.entries[i][1]) {
                     return true;
                 }
             }
             return false;
+        },
+
+        hashCode: function () {
+            return hashObject(this.entries[0][0]);
+        },
+
+        get: function (key) {
+            var entry = this.getEntryForKey(key);
+            return entry ? entry[1] : null;
         }
     };
 
-    /*----------------------------------------------------------------------------------------------------------------*/
-
-    // Supporting functions for searching hashtable buckets
-
-    function searchBuckets(buckets, hash)
-    {
-        var i = buckets.length, bucket;
-        while (i--)
-        {
-            bucket = buckets[i];
-            if (hash === bucket[0])
-            {
+    function searchBuckets(buckets, hash) {
+        for (var i = 0, len = buckets.length; i < len; ++i) {
+            if (buckets[i][0] === hash) {
                 return i;
             }
         }
         return null;
     }
 
-    function getBucketForHash(bucketsByHash, hash)
-    {
-        var bucket = bucketsByHash[hash];
-
-        // Check that this is a genuine bucket and not something inherited from the bucketsByHash's prototype
-        return ( bucket && (bucket instanceof Bucket) ) ? bucket : null;
+    function getBucketForHash(bucketsByHash, hash) {
+        return bucketsByHash[hash] || null;
     }
 
-    /*----------------------------------------------------------------------------------------------------------------*/
-
-    function Hashtable(hashingFunctionParam, equalityFunctionParam)
-    {
-        var that = this;
+    function Hashtable(hashingFunctionParam, equalityFunctionParam) {
         var buckets = [];
         var bucketsByHash = {};
 
-        var hashingFunction = (typeof hashingFunctionParam == FUNCTION) ? hashingFunctionParam : hashObject;
-        var equalityFunction = (typeof equalityFunctionParam == FUNCTION) ? equalityFunctionParam : null;
+        var hashingFunction = hashingFunctionParam || hashObject;
+        var equalityFunction = equalityFunctionParam || null;
 
-        this.put = function (key, value)
-        {
-            checkKey(key);
-            checkValue(value);
+        this.put = function (key, value) {
+            if (key === null || typeof key === 'undefined') {
+                throw new Error("null is not a valid key");
+            }
+            if (value === null || typeof value === 'undefined') {
+                throw new Error("value must not be undefined");
+            }
             var hash = hashingFunction(key), bucket, bucketEntry, oldValue = null;
 
-            // Check if a bucket exists for the bucket key
             bucket = getBucketForHash(bucketsByHash, hash);
-            if (bucket)
-            {
-                // Check this bucket to see if it already contains this key
+            if (bucket) {
                 bucketEntry = bucket.getEntryForKey(key);
-                if (bucketEntry)
-                {
-                    // This bucket entry is the current mapping of key to value, so replace old value and we're done.
+                if (bucketEntry) {
                     oldValue = bucketEntry[1];
                     bucketEntry[1] = value;
-                } else
-                {
-                    // The bucket does not contain an entry for this key, so add one
+                } else {
                     bucket.addEntry(key, value);
                 }
-            } else
-            {
-                // No bucket exists for the key, so create one and put our key/value mapping in
+            } else {
                 bucket = new Bucket(hash, key, value, equalityFunction);
                 buckets[buckets.length] = bucket;
                 bucketsByHash[hash] = bucket;
@@ -298,70 +210,53 @@ gamecore.Hashtable = (function ()
             return oldValue;
         };
 
-        this.get = function (key)
-        {
-            checkKey(key);
-
+        this.get = function (key) {
+            if (key === null || typeof key === 'undefined') {
+                throw new Error("null is not a valid key");
+            }
             var hash = hashingFunction(key);
-
-            // Check if a bucket exists for the bucket key
             var bucket = getBucketForHash(bucketsByHash, hash);
-            if (bucket)
-            {
-                // Check this bucket to see if it contains this key
+            if (bucket) {
                 var bucketEntry = bucket.getEntryForKey(key);
-                if (bucketEntry)
-                {
-                    // This bucket entry is the current mapping of key to value, so return the value.
+                if (bucketEntry) {
                     return bucketEntry[1];
                 }
             }
             return null;
         };
 
-        this.containsKey = function (key)
-        {
-            checkKey(key);
-            var bucketKey = hashingFunction(key);
-
-            // Check if a bucket exists for the bucket key
-            var bucket = getBucketForHash(bucketsByHash, bucketKey);
-
+        this.containsKey = function (key) {
+            if (key === null || typeof key === 'undefined') {
+                throw new Error("null is not a valid key");
+            }
+            var hash = hashingFunction(key);
+            var bucket = getBucketForHash(bucketsByHash, hash);
             return bucket ? bucket.containsKey(key) : false;
         };
 
-        this.containsValue = function (value)
-        {
-            checkValue(value);
+        this.containsValue = function (value) {
             var i = buckets.length;
-            while (i--)
-            {
-                if (buckets[i].containsValue(value))
-                {
+            while (i--) {
+                if (buckets[i].containsValue(value)) {
                     return true;
                 }
             }
             return false;
         };
 
-        this.clear = function ()
-        {
+        this.clear = function () {
             buckets.length = 0;
             bucketsByHash = {};
         };
 
-        this.isEmpty = function ()
-        {
+        this.isEmpty = function () {
             return !buckets.length;
         };
 
-        var createBucketAggregator = function (bucketFuncName)
-        {
-            return function ()
-            {
+        var createBucketAggregator = function (bucketFuncName) {
+            return function () {
                 var aggregated = [], i = buckets.length;
-                while (i--)
-                {
+                while (i--) {
                     buckets[i][bucketFuncName](aggregated);
                 }
                 return aggregated;
@@ -372,90 +267,71 @@ gamecore.Hashtable = (function ()
         this.values = createBucketAggregator("values");
         this.entries = createBucketAggregator("getEntries");
 
-        this.remove = function (key)
-        {
-            checkKey(key);
+        this.remove = function (key) {
+            if (key === null || typeof key === 'undefined') {
+                throw new Error("null is not a valid key");
+            }
+            var hash = hashingFunction(key);
+            var bucketIndex, oldValue = null;
 
-            var hash = hashingFunction(key), bucketIndex, oldValue = null;
-
-            // Check if a bucket exists for the bucket key
             var bucket = getBucketForHash(bucketsByHash, hash);
 
-            if (bucket)
-            {
-                // Remove entry from this bucket for this key
+            if (bucket) {
                 oldValue = bucket.removeEntryForKey(key);
-                if (oldValue !== null)
-                {
-                    // Entry was removed, so check if bucket is empty
-                    if (!bucket.entries.length)
-                    {
-                        // Bucket is empty, so remove it from the bucket collections
-                        bucketIndex = searchBuckets(buckets, hash);
-                        arrayRemoveAt(buckets, bucketIndex);
-                        delete bucketsByHash[hash];
-                    }
+                if (!bucket.entries.length) {
+                    bucketIndex = searchBuckets(buckets, hash);
+                    buckets.splice(bucketIndex, 1);
+                    delete bucketsByHash[hash];
                 }
             }
             return oldValue;
         };
 
-        this.size = function ()
-        {
+        this.size = function () {
             var total = 0, i = buckets.length;
-            while (i--)
-            {
+            while (i--) {
                 total += buckets[i].entries.length;
             }
             return total;
         };
 
-        this.each = function (callback)
-        {
-            var entries = that.entries(), i = entries.length, entry;
-            while (i--)
-            {
+        this.each = function (callback) {
+            var entries = this.entries(), i = entries.length, entry;
+            while (i--) {
                 entry = entries[i];
                 callback(entry[0], entry[1]);
             }
         };
 
-        this.putAll = function (hashtable, conflictCallback)
-        {
+        this.putAll = function (hashtable, conflictCallback) {
             var entries = hashtable.entries();
             var entry, key, value, thisValue, i = entries.length;
-            var hasConflictCallback = (typeof conflictCallback == FUNCTION);
-            while (i--)
-            {
+            var hasConflictCallback = typeof conflictCallback === 'function';
+            while (i--) {
                 entry = entries[i];
                 key = entry[0];
                 value = entry[1];
 
-                // Check for a conflict. The default behaviour is to overwrite the value for an existing key
-                if (hasConflictCallback && (thisValue = that.get(key)))
-                {
+                if (hasConflictCallback && (thisValue = this.get(key))) {
                     value = conflictCallback(key, thisValue, value);
                 }
-                that.put(key, value);
+                this.put(key, value);
             }
         };
 
-        this.clone = function ()
-        {
+        this.clone = function () {
             var clone = new Hashtable(hashingFunctionParam, equalityFunctionParam);
-            clone.putAll(that);
+            clone.putAll(this);
             return clone;
         };
 
         /**
          * Added by martin@playcratlabs.com to support debug dumping of hash arrays
          */
-        this.toString = function ()
-        {
+        this.toString = function () {
             var result = '';
             var keys = this.keys();
-            for (var i = 0; i < keys.length; i++)
-            {
+            for (var i = 0; i < keys.length; i++) {
                 var obj = this.get(keys[i]);
                 result += keys[i].toString() + ' = ' + obj.toString() + '\n';
             }
@@ -464,5 +340,5 @@ gamecore.Hashtable = (function ()
         }
     }
 
-    return Hashtable;
+    gamecore.Hashtable = Hashtable;
 })();
